refactor: move the types folder from app to src
